Implement Firebase functionality with addExpense action
Add Redux-Thunk to the Redux actions in order to be able to use functions as actions instead of only objects.
Reformat the code in order to accomodate for the new change, as the application now uses props.startAddExpense as a method to
dispatch the addExpense action to the redux-store. And the store now waits until the Firebase has successfully saved the
expense to the database THEN returns the id of the saved expense (which is referenced as 'ref' within the function) and allows to dispatch the original addExpense() action to commit the change to the store.